Continue adapting the simulation to the response-scale analyses (of
the latent projection): Now adapting section "Post-processing".

Thereby also remove the commented `plotter_single()` function because currently,
it can't be used since returning the output of `plot.vsel()` leads to memory
issues when running the simulation in parallel (at least on a local machine; I
haven't tried it on a cluster yet). If the `plotter_single()` function were not
removed, it would also have to be adapted to the response-scale analyses of the
latent projection.

Also update a comment concerning the former branch `elpd4`.