Remove several plots:
1. The `_eb` plots (containing the error bars for a subset of all simulation
iterations). In principle, they could be helpful, but are not needed right now.
And if they are used in the future, then the y-axis labels should include
something like "(\pm SE)". Furthermore, it might make sense to "dodge" the error
bars within a given submodel size (i.e., for the different simulation
iterations) and to decrease the width of the horizontal bars at the top and the
bottom so that it's easier to distinguish them.
2. The separate SE plots (`_se`, but not `_diff_se`). In principle, they could
be helpful, but are not needed right now. And if they are used in the future,
then the y-axis limits should be the same for the augmented-data and the latent
plot.
3. The `_zoom`-ed plots. In the manuscript, the `_full` plots currently make
more sense.