Add commented code for incorporating the reference model statistic
into the second y-axis. However, this is not a good idea because the reference
model statistic varies from simulation iteration to simulation iteration, so all
we could do would be to add some aggregated reference model statistic, but this
does not match the individual reference model statistics from the simulation
iterations.