Major overhaul of the analysis of the predictive performance at
the smaller one of the 2 suggested sizes. In particular, don't select the
"extreme" simulation iterations only based on the GMPD difference anymore (and
not only the minimum and maximum anymore, but the 3 smallest and 3 largest GMPD
differences), but also---primarily---based on the MLPD difference (for
consistency with the `_maxdiff` objects). Also don't create the individual
absolute-scale MLPD plots for 21 randomly chosen simulation iterations anymore
(because we now have individual absolute-scale MLPD plots for the 6 most extreme
simulation iterations in terms of MLPD difference and---secondarily---in terms
of GMPD difference).